refactor(query): split lightweight grammar metadata
Problem: #2006/#1844 need query completion and builder surfaces to consume grammar metadata without importing the parser/lowerer module. The iter target-state patches identified this, and the current code still had completions pulling EXPRESSION_FIELD_REGISTRY and structural registries from expression.py.\n\nWhat changed: move field, structural-unit, count, and date query metadata into polylogue.archive.query.metadata. expression.py imports and re-exports those names so existing parser callsites keep using the canonical Lark compiler, while completions and shell completion import the lightweight metadata module directly. Add an import-boundary test proving query completions do not load polylogue.archive.query.expression.\n\nVerification: nix develop --command devtools test tests/unit/archive/test_query_metadata.py tests/unit/cli/test_command_aux_runtime.py -k 'query_field_candidates or query_structural_unit_candidates or query_completions_do_not_import_expression_parser' (5 passed); nix develop --command devtools test tests/unit/cli/test_query_expression.py -k 'JsonSpecStrictMode or FIELD_DESCRIPTOR or query_field or structural or count_query or date_query' (11 passed); combined focused query slice (18 passed); nix develop --command devtools verify --quick (exit_code 0).\n\nRef #2006\nRef #1844
