+++ docs/topology-status.md
@@ -19,7 +19,7 @@ Generated by `devtools render-topology-status`. Reads `docs/plans/topology-targe
 | archive-phase | — | 2 | 2 | 0 | 0 |
 | archive-projection | — | 5 | 5 | 0 | 0 |
 | archive-provider | — | 2 | 2 | 0 | 0 |
-| archive-query | archive query semantics | 24 | 24 | 0 | 0 |
+| archive-query | archive query semantics | 25 | 25 | 0 | 0 |
 | archive-raw-payload | — | 5 | 5 | 0 | 0 |
 | archive-semantic | — | 13 | 13 | 0 | 0 |
 | archive-session | — | 16 | 16 | 0 | 0 |
@@ -32,6 +32,6 @@ Generated by `devtools render-topology-status`. Reads `docs/plans/topology-targe
 - **Kernel** (polylogue/ root): 10
 - **Primitives** (storage-root): 16
 - **TBD** (cell needs explicit assignment): 0
-- **Total declared**: 838
-- **Realized polylogue/**/*.py**: 838 files declared
+- **Total declared**: 839
+- **Realized polylogue/**/*.py**: 839 files declared
 <!-- END GENERATED: topology-status -->
